Add Phased Array and Linear Probe Geometries (#34)
## Description
This PR adds the ability to simulation linear and curvilinear probe
geometries. This is done by creating a `BaseProbe` class and deriving
the classes for all probe geometries. In this proposal, every probe
geometry has it's own scan conversion method to render the B-mode image.
Future work would unify the scan conversion methods to a unified
(simple) method to reduce conditional logic.

The `server.py` example was updated to allow for the dynamic selection
of ultrasound probe in the browser.

---------

Co-authored-by: Mingxin Zheng <18563433+mingxin-zheng@users.noreply.github.com>